Network.AWS.ImportExport

Description

AWS Import/Export accelerates moving large amounts of data into and out of AWS using portable storage devices for transport. AWS transfers your data directly onto and off of storage devices using Amazon’s high-speed internal network and bypassing the Internet. For significant data sets, AWS Import/Export is often faster than Internet transfer and more cost effective than upgrading your connectivity.

Documentation

module Network.AWS.ImportExport.CancelJob

module Network.AWS.ImportExport.CreateJob

module Network.AWS.ImportExport.GetShippingLabel

module Network.AWS.ImportExport.GetStatus

module Network.AWS.ImportExport.ListJobs

module Network.AWS.ImportExport.Types

module Network.AWS.ImportExport.UpdateJob