Network.AWS.ImportExport.CancelJob

Description

This operation cancels a specified job. Only the job owner can cancel it. The operation fails if the job has already started or is complete.

Creating a Request

cancelJob

Arguments

:: Text
cJobId

-> CancelJob

Creates a value of CancelJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cAPIVersion - Undocumented member.

cJobId - Undocumented member.

data CancelJob

Input structure for the CancelJob operation.

See: cancelJob smart constructor.

Request Lenses

cAPIVersion :: Lens' CancelJob (Maybe Text)

Undocumented member.

cJobId :: Lens' CancelJob Text

Undocumented member.

Destructuring the Response

cancelJobResponse

Arguments

:: Int
crsResponseStatus

-> CancelJobResponse

Creates a value of CancelJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsSuccess - Undocumented member.

crsResponseStatus - -- | The response status code.

data CancelJobResponse

Output structure for the CancelJob operation.

See: cancelJobResponse smart constructor.

Response Lenses

crsSuccess :: Lens' CancelJobResponse (Maybe Bool)

Undocumented member.

crsResponseStatus :: Lens' CancelJobResponse Int

- | The response status code.