Network.AWS.ImportExport.CreateJob

Description

This operation initiates the process of scheduling an upload or download of your data. You include in the request a manifest that describes the data transfer specifics. The response to the request includes a job ID, which you can use in other operations, a signature that you use to identify your storage device, and the address where you should ship your storage device.

Request

data CreateJob

Input structure for the CreateJob operation.

Request constructor

mkCreateJob

Arguments

:: JobType
cj1JobType

-> Text
cj1Manifest

-> Bool
cj1ValidateOnly

-> CreateJob

Smart constructor for the minimum required parameters to construct a valid CreateJob request.

The fields accessible through corresponding lenses are:

JobType :: JobType

Manifest :: Text

ManifestAddendum :: Maybe Text

ValidateOnly :: Bool

Request lenses

cj1JobType :: Lens' CreateJob JobType

Specifies whether the job to initiate is an import or export job.

cj1Manifest :: Lens' CreateJob Text

The UTF-8 encoded text of the manifest file.

cj1ManifestAddendum :: Lens' CreateJob (Maybe Text)

For internal use only.

cj1ValidateOnly :: Lens' CreateJob Bool

Validate the manifest and parameter values in the request but do not actually create a job.

Response

data CreateJobResponse

Output structure for the CreateJob operation.

Response constructor

mkCreateJobResponse :: CreateJobResponse

Smart constructor for the minimum required parameters to construct a valid CreateJobResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

JobId :: Maybe Text

JobType :: Maybe JobType

AwsShippingAddress :: Maybe Text

Signature :: Maybe Text

SignatureFileContents :: Maybe Text

WarningMessage :: Maybe Text

Response lenses

cjrrJobId :: Lens' CreateJobResponse (Maybe Text)

A unique identifier which refers to a particular job.

cjrrJobType :: Lens' CreateJobResponse (Maybe JobType)

Specifies whether the job to initiate is an import or export job.

cjrrAwsShippingAddress :: Lens' CreateJobResponse (Maybe Text)

Address you ship your storage device to.

cjrrSignature :: Lens' CreateJobResponse (Maybe Text)

An encrypted code used to authenticate the request and response, for example, "DV+TpDfx1TdSE9ktyK9kbDTVI=". Only use this value is you want to create the signature file yourself. Generally you should use the SignatureFileContents value.

cjrrSignatureFileContents :: Lens' CreateJobResponse (Maybe Text)

The actual text of the SIGNATURE file to be written to disk.

cjrrWarningMessage :: Lens' CreateJobResponse (Maybe Text)

An optional message notifying you of non-fatal issues with the job, such as use of an incompatible Amazon S3 bucket name.