Network.AWS.ImportExport.GetShippingLabel

Description

This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job. You can only return information about jobs you own.

http://docs.aws.amazon.com/AWSImportExport/latest/DG/WebGetShippingLabel.html

Request

data GetShippingLabel

Request constructor

getShippingLabel :: GetShippingLabel

GetShippingLabel constructor.

The fields accessible through corresponding lenses are:

gslAPIVersion :: Maybe Text

gslCity :: Maybe Text

gslCompany :: Maybe Text

gslCountry :: Maybe Text

gslJobIds :: [Text]

gslName :: Maybe Text

gslPhoneNumber :: Maybe Text

gslPostalCode :: Maybe Text

gslStateOrProvince :: Maybe Text

gslStreet1 :: Maybe Text

gslStreet2 :: Maybe Text

gslStreet3 :: Maybe Text

Request lenses

gslAPIVersion :: Lens' GetShippingLabel (Maybe Text)

gslCity :: Lens' GetShippingLabel (Maybe Text)

gslCompany :: Lens' GetShippingLabel (Maybe Text)

gslCountry :: Lens' GetShippingLabel (Maybe Text)

gslJobIds :: Lens' GetShippingLabel [Text]

gslName :: Lens' GetShippingLabel (Maybe Text)

gslPhoneNumber :: Lens' GetShippingLabel (Maybe Text)

gslPostalCode :: Lens' GetShippingLabel (Maybe Text)

gslStateOrProvince :: Lens' GetShippingLabel (Maybe Text)

gslStreet1 :: Lens' GetShippingLabel (Maybe Text)

gslStreet2 :: Lens' GetShippingLabel (Maybe Text)

gslStreet3 :: Lens' GetShippingLabel (Maybe Text)

Response

data GetShippingLabelResponse

Response constructor

getShippingLabelResponse :: GetShippingLabelResponse

GetShippingLabelResponse constructor.

The fields accessible through corresponding lenses are:

gslrShippingLabelURL :: Maybe Text

gslrWarning :: Maybe Text

Response lenses

gslrShippingLabelURL :: Lens' GetShippingLabelResponse (Maybe Text)

gslrWarning :: Lens' GetShippingLabelResponse (Maybe Text)