Network.AWS.ImportExport.GetStatus

Description

This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job. You can only return information about jobs you own.

http://docs.aws.amazon.com/AWSImportExport/latest/DG/WebGetStatus.html

Request

data GetStatus

Request constructor

getStatus

Arguments

:: Text
gsJobId

-> GetStatus

GetStatus constructor.

The fields accessible through corresponding lenses are:

gsAPIVersion :: Maybe Text

gsJobId :: Text

Request lenses

gsAPIVersion :: Lens' GetStatus (Maybe Text)

gsJobId :: Lens' GetStatus Text

Response

data GetStatusResponse

Response constructor

getStatusResponse :: GetStatusResponse

GetStatusResponse constructor.

The fields accessible through corresponding lenses are:

gsrArtifactList :: [Artifact]

gsrCarrier :: Maybe Text

gsrCreationDate :: Maybe UTCTime

gsrCurrentManifest :: Maybe Text

gsrErrorCount :: Maybe Int

gsrJobId :: Maybe Text

gsrJobType :: Maybe JobType

gsrLocationCode :: Maybe Text

gsrLocationMessage :: Maybe Text

gsrLogBucket :: Maybe Text

gsrLogKey :: Maybe Text

gsrProgressCode :: Maybe Text

gsrProgressMessage :: Maybe Text

gsrSignature :: Maybe Text

gsrSignatureFileContents :: Maybe Text

gsrTrackingNumber :: Maybe Text

Response lenses

gsrArtifactList :: Lens' GetStatusResponse [Artifact]

gsrCarrier :: Lens' GetStatusResponse (Maybe Text)

gsrCreationDate :: Lens' GetStatusResponse (Maybe UTCTime)

gsrCurrentManifest :: Lens' GetStatusResponse (Maybe Text)

gsrErrorCount :: Lens' GetStatusResponse (Maybe Int)

gsrJobId :: Lens' GetStatusResponse (Maybe Text)

gsrJobType :: Lens' GetStatusResponse (Maybe JobType)

gsrLocationCode :: Lens' GetStatusResponse (Maybe Text)

gsrLocationMessage :: Lens' GetStatusResponse (Maybe Text)

gsrLogBucket :: Lens' GetStatusResponse (Maybe Text)

gsrLogKey :: Lens' GetStatusResponse (Maybe Text)

gsrProgressCode :: Lens' GetStatusResponse (Maybe Text)

gsrProgressMessage :: Lens' GetStatusResponse (Maybe Text)

gsrSignature :: Lens' GetStatusResponse (Maybe Text)

gsrSignatureFileContents :: Lens' GetStatusResponse (Maybe Text)

gsrTrackingNumber :: Lens' GetStatusResponse (Maybe Text)