Network.AWS.ImportExport.ListJobs

Description

This operation returns the jobs associated with the requester. AWS Import/Export lists the jobs in reverse chronological order based on the date of creation. For example if Job Test1 was created 2009Dec30 and Test2 was created 2010Feb05, the ListJobs operation would return Test2 followed by Test1.

http://docs.aws.amazon.com/AWSImportExport/latest/DG/WebListJobs.html

Request

data ListJobs

Request constructor

listJobs :: ListJobs

ListJobs constructor.

The fields accessible through corresponding lenses are:

ljAPIVersion :: Maybe Text

ljMarker :: Maybe Text

ljMaxJobs :: Maybe Int

Request lenses

ljAPIVersion :: Lens' ListJobs (Maybe Text)

ljMarker :: Lens' ListJobs (Maybe Text)

ljMaxJobs :: Lens' ListJobs (Maybe Int)

Response

data ListJobsResponse

Response constructor

listJobsResponse :: ListJobsResponse

ListJobsResponse constructor.

The fields accessible through corresponding lenses are:

ljrIsTruncated :: Maybe Bool

ljrJobs :: [Job]

Response lenses

ljrIsTruncated :: Lens' ListJobsResponse (Maybe Bool)

ljrJobs :: Lens' ListJobsResponse [Job]