Network.AWS.ImportExport.Monadic

Description

This module is provided for convenience. It offers an alternative to the common idiom of supplying required fields to an operations's smart constructor, using the operation's lenses to modify additional fields, and then sending or paginating the request.

As an example: using Network.AWS.ImportExport with the smart constructor and basic lens syntax, before explicitly calling send:

import Control.Monad.Trans.AWS
import Network.AWS.ImportExport

send $ (mkOperationName w x)
     & onLensField1 .~ y
     & onLensField2 .~ z

Versus using Network.AWS.ImportExport.Monadic with the State operator variants from Control.Lens.Setter such as .= to modify any additional request parameters before sending:

import Control.Applicative
import Network.AWS.ImportExport.Monadic

operationName w x $ do
    onLensField1 .= y
    onLensField2 .= z

-- Or to void any additional parameters outside of those required using return ():
operationName w x $ return ()

CancelJob

This operation cancels a specified job. Only the job owner can cancel it. The operation fails if the job has already started or is complete.

See: CancelJob

cancelJob

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cjJobId

-> m CancelJobResponse

cancelJobCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cjJobId

-> m (Either ServiceEr CancelJobResponse)

CreateJob

This operation initiates the process of scheduling an upload or download of your data. You include in the request a manifest that describes the data transfer specifics. The response to the request includes a job ID, which you can use in other operations, a signature that you use to identify your storage device, and the address where you should ship your storage device.

See: CreateJob

createJob

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> JobType
cj1JobType

-> Text
cj1Manifest

-> Bool
cj1ValidateOnly

-> State CreateJob a

-> m CreateJobResponse

createJobCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> JobType
cj1JobType

-> Text
cj1Manifest

-> Bool
cj1ValidateOnly

-> State CreateJob a

-> m (Either ServiceEr CreateJobResponse)

GetStatus

This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job. You can only return information about jobs you own.

See: GetStatus

getStatus

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gsJobId

-> m GetStatusResponse

getStatusCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gsJobId

-> m (Either ServiceEr GetStatusResponse)

ListJobs

This operation returns the jobs associated with the requester. AWS Import/Export lists the jobs in reverse chronological order based on the date of creation. For example if Job Test1 was created 2009Dec30 and Test2 was created 2010Feb05, the ListJobs operation would return Test2 followed by Test1.

See: ListJobs

listJobs :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State ListJobs a -> Source m ListJobsResponse

listJobsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State ListJobs a -> Source m (Either ServiceEr ListJobsResponse)

UpdateJob

You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file. The manifest file attached to this request replaces the original manifest file. You can only use the operation after a CreateJob request but before the data transfer starts and you can only use it on jobs you own.

See: UpdateJob

updateJob

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ujJobId

-> Text
ujManifest

-> JobType
ujJobType

-> Bool
ujValidateOnly

-> m UpdateJobResponse

updateJobCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ujJobId

-> Text
ujManifest

-> JobType
ujJobType

-> Bool
ujValidateOnly

-> m (Either ServiceEr UpdateJobResponse)

Re-exported

module Network.AWS.ImportExport

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()