Network.AWS.ImportExport.Types

Description

AWS Import/Export accelerates moving large amounts of data into and out of AWS using portable storage devices for transport. AWS transfers your data directly onto and off of storage devices using Amazon’s high-speed internal network and bypassing the Internet. For significant data sets, AWS Import/Export is often faster than Internet transfer and more cost effective than upgrading your connectivity.

Service

data ImportExport

Supported version (2010-06-01) of the AWS Import/Export service.

Errors

data ImportExportError

A sum type representing possible errors returned by the ImportExport service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

BucketPermissionException
The account specified does not have the appropriate bucket permissions.

Fields

_bpeMessage :: Maybe Text

CanceledJobIdException
The specified job ID has been canceled and is no longer valid.

Fields

_cjieMessage :: Maybe Text

ExpiredJobIdException
Indicates that the specified job has expired out of the system.

Fields

_ejieMessage :: Maybe Text

ImportExportClient HttpException

ImportExportSerializer String

ImportExportService String

InvalidAccessKeyIdException
The AWS Access Key ID specified in the request did not match the manifest's accessKeyId value. The manifest and the request authentication must use the same AWS Access Key ID.

Fields

_iakieMessage :: Maybe Text

InvalidAddressException
The address specified in the manifest is invalid.

Fields

_iaeMessage :: Maybe Text

InvalidCustomsException
One or more customs parameters was invalid. Please correct and resubmit.

Fields

_iceMessage :: Maybe Text

InvalidFileSystemException
File system specified in export manifest is invalid.

Fields

_ifseMessage :: Maybe Text

InvalidJobIdException
The JOBID was missing, not found, or not associated with the AWS account.

Fields

_ijieMessage :: Maybe Text

InvalidManifestFieldException
One or more manifest fields was invalid. Please correct and resubmit.

Fields

_imfeMessage :: Maybe Text

InvalidParameterException
One or more parameters had an invalid value.

Fields

_ipeMessage :: Maybe Text

MalformedManifestException
Your manifest is not well-formed.

Fields

_mmeMessage :: Maybe Text

MissingCustomsException
One or more required customs parameters was missing from the manifest.

Fields

_mceMessage :: Maybe Text

MissingManifestFieldException
One or more required fields were missing from the manifest file. Please correct and resubmit.

Fields

_mmfeMessage :: Maybe Text

MissingParameterException
One or more required parameters was missing from the request.

Fields

_mpeMessage :: Maybe Text

MultipleRegionsException
Your manifest file contained buckets from multiple regions. A job is restricted to buckets from one region. Please correct and resubmit.

Fields

_mreMessage :: Maybe Text

NoSuchBucketException
The specified bucket does not exist. Create the specified bucket or change the manifest's bucket, exportBucket, or logBucket field to a bucket that the account, as specified by the manifest's Access Key ID, has write permissions to.

Fields

_nsbeMessage :: Maybe Text

UnableToCancelJobIdException
AWS Import/Export cannot cancel the job.

Fields

_utcjieMessage :: Maybe Text

_BucketPermissionException :: Prism' ImportExportError (Maybe Text)

The account specified does not have the appropriate bucket permissions.

See: BucketPermissionException

_CanceledJobIdException :: Prism' ImportExportError (Maybe Text)

The specified job ID has been canceled and is no longer valid.

See: CanceledJobIdException

_ExpiredJobIdException :: Prism' ImportExportError (Maybe Text)

Indicates that the specified job has expired out of the system.

See: ExpiredJobIdException

_ImportExportClient :: Prism' ImportExportError HttpException

See: ImportExportClient

_ImportExportSerializer :: Prism' ImportExportError String

See: ImportExportSerializer

_ImportExportService :: Prism' ImportExportError String

See: ImportExportService

_InvalidAccessKeyIdException :: Prism' ImportExportError (Maybe Text)

The AWS Access Key ID specified in the request did not match the manifest's accessKeyId value. The manifest and the request authentication must use the same AWS Access Key ID.

See: InvalidAccessKeyIdException

_InvalidAddressException :: Prism' ImportExportError (Maybe Text)

The address specified in the manifest is invalid.

See: InvalidAddressException

_InvalidCustomsException :: Prism' ImportExportError (Maybe Text)

One or more customs parameters was invalid. Please correct and resubmit.

See: InvalidCustomsException

_InvalidFileSystemException :: Prism' ImportExportError (Maybe Text)

File system specified in export manifest is invalid.

See: InvalidFileSystemException

_InvalidJobIdException :: Prism' ImportExportError (Maybe Text)

The JOBID was missing, not found, or not associated with the AWS account.

See: InvalidJobIdException

_InvalidManifestFieldException :: Prism' ImportExportError (Maybe Text)

One or more manifest fields was invalid. Please correct and resubmit.

See: InvalidManifestFieldException

_InvalidParameterException :: Prism' ImportExportError (Maybe Text)

One or more parameters had an invalid value.

See: InvalidParameterException

_MalformedManifestException :: Prism' ImportExportError (Maybe Text)

Your manifest is not well-formed.

See: MalformedManifestException

_MissingCustomsException :: Prism' ImportExportError (Maybe Text)

One or more required customs parameters was missing from the manifest.

See: MissingCustomsException

_MissingManifestFieldException :: Prism' ImportExportError (Maybe Text)

One or more required fields were missing from the manifest file. Please correct and resubmit.

See: MissingManifestFieldException

_MissingParameterException :: Prism' ImportExportError (Maybe Text)

One or more required parameters was missing from the request.

See: MissingParameterException

_MultipleRegionsException :: Prism' ImportExportError (Maybe Text)

Your manifest file contained buckets from multiple regions. A job is restricted to buckets from one region. Please correct and resubmit.

See: MultipleRegionsException

_NoSuchBucketException :: Prism' ImportExportError (Maybe Text)

The specified bucket does not exist. Create the specified bucket or change the manifest's bucket, exportBucket, or logBucket field to a bucket that the account, as specified by the manifest's Access Key ID, has write permissions to.

See: NoSuchBucketException

_UnableToCancelJobIdException :: Prism' ImportExportError (Maybe Text)

AWS Import/Export cannot cancel the job.

See: UnableToCancelJobIdException

XML

xmlOptions :: Tagged a XMLOptions

JobType

data JobType

Constructors

JobTypeExport
Export

JobTypeImport
Import

Job

data Job

Representation of a job returned by the ListJobs operation.

job

Arguments

:: Text
jJobId

-> ISO8601
jCreationDate

-> Bool
jIsCanceled

-> JobType
jJobType

-> Job

Smart constructor for the minimum required fields to construct a valid Job data type.

Job is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

JobId :: Text

CreationDate :: ISO8601

IsCanceled :: Bool

JobType :: JobType

jJobId :: Lens' Job Text

A unique identifier which refers to a particular job.

jCreationDate :: Lens' Job ISO8601

Timestamp of the CreateJob request in ISO8601 date format. For example "2010-03-28T20:27:35Z".

jIsCanceled :: Lens' Job Bool

Indicates whether the job was canceled.

jJobType :: Lens' Job JobType

Specifies whether the job to initiate is an import or export job.