Network.AWS.ImportExport.V2010_06_01

Description

AWS Import/Export accelerates moving large amounts of data into and out of AWS using portable storage devices for transport. AWS transfers your data directly onto and off of storage devices using Amazon’s high-speed internal network and bypassing the Internet. For significant data sets, AWS Import/Export is often faster than Internet transfer and more cost effective than upgrading your connectivity.

Documentation

data CancelJob

Constructors

CancelJob

Fields

_cjjJobId :: Text

A unique identifier which refers to a particular job.

cjjJobId :: Iso' CancelJob Text

data CancelJobResponse

Constructors

CancelJobResponse

Fields

_cjpSuccess :: Maybe Bool

Specifies whether (true) or not (false) AWS Import/Export updated your job.

cjpSuccess :: Iso' CancelJobResponse (Maybe Bool)

createJob

Arguments

:: JobType
_cjiJobType

-> Text
_cjiManifest

-> Bool
_cjiValidateOnly

-> CreateJob

Minimum specification for a CreateJob request.

data CreateJob

Constructors

CreateJob

Fields

_cjiJobType :: JobType

Specifies whether the job to initiate is an import or export job.

_cjiManifest :: Text

The UTF-8 encoded text of the manifest file.

_cjiValidateOnly :: Bool

Validate the manifest and parameter values in the request but do not actually create a job.

_cjiManifestAddendum :: Maybe Text

For internal use only.

cjiValidateOnly :: Lens' CreateJob Bool

cjiManifestAddendum :: Lens' CreateJob (Maybe Text)

cjiManifest :: Lens' CreateJob Text

cjiJobType :: Lens' CreateJob JobType

data CreateJobResponse

Constructors

CreateJobResponse

Fields

_cjoAwsShippingAddress :: Maybe Text

Address you ship your storage device to.

_cjoJobId :: Maybe Text

A unique identifier which refers to a particular job.

_cjoJobType :: Maybe JobType

Specifies whether the job to initiate is an import or export job.

_cjoSignature :: Maybe Text

An encrypted code used to authenticate the request and response, for example, "DV+TpDfx1TdSE9ktyK9kbDTVI=". Only use this value is you want to create the signature file yourself. Generally you should use the SignatureFileContents value.

_cjoSignatureFileContents :: Maybe Text

The actual text of the SIGNATURE file to be written to disk.

_cjoWarningMessage :: Maybe Text

An optional message notifying you of non-fatal issues with the job, such as use of an incompatible Amazon S3 bucket name.

cjoWarningMessage :: Lens' CreateJobResponse (Maybe Text)

cjoSignatureFileContents :: Lens' CreateJobResponse (Maybe Text)

cjoSignature :: Lens' CreateJobResponse (Maybe Text)

cjoJobType :: Lens' CreateJobResponse (Maybe JobType)

cjoJobId :: Lens' CreateJobResponse (Maybe Text)

cjoAwsShippingAddress :: Lens' CreateJobResponse (Maybe Text)

data GetStatus

Constructors

GetStatus

Fields

_gsiJobId :: Text

A unique identifier which refers to a particular job.

gsiJobId :: Iso' GetStatus Text

data GetStatusResponse

Constructors

GetStatusResponse

Fields

_gsoAwsShippingAddress :: Maybe Text

Address you ship your storage device to.

_gsoCarrier :: Maybe Text

Name of the shipping company. This value is included when the LocationCode is Returned.

_gsoCreationDate :: Maybe ISO8601

Timestamp of the CreateJob request in ISO8601 date format. For example "2010-03-28T20:27:35Z".

_gsoCurrentManifest :: Maybe Text

The last manifest submitted, which will be used to process the job.

_gsoErrorCount :: Maybe Integer

Number of errors. We return this value when the ProgressCode is Success or SuccessWithErrors.

_gsoJobId :: Maybe Text

A unique identifier which refers to a particular job.

_gsoJobType :: Maybe JobType

Specifies whether the job to initiate is an import or export job.

_gsoLocationCode :: Maybe Text

A token representing the location of the storage device, such as AtAWS.

_gsoLocationMessage :: Maybe Text

A more human readable form of the physical location of the storage device.

_gsoLogBucket :: Maybe Text

Amazon S3 bucket for user logs.

_gsoLogKey :: Maybe Text

The key where the user logs were stored.

_gsoProgressCode :: Maybe Text

A token representing the state of the job, such as Started.

_gsoProgressMessage :: Maybe Text

A more human readable form of the job status.

_gsoSignature :: Maybe Text

An encrypted code used to authenticate the request and response, for example, "DV+TpDfx1TdSE9ktyK9kbDTVI=". Only use this value is you want to create the signature file yourself. Generally you should use the SignatureFileContents value.

_gsoSignatureFileContents :: Maybe Text

An encrypted code used to authenticate the request and response, for example, "DV+TpDfx1TdSE9ktyK9kbDTVI=". Only use this value is you want to create the signature file yourself. Generally you should use the SignatureFileContents value.

_gsoTrackingNumber :: Maybe Text

The shipping tracking number assigned by AWS Import/Export to the storage device when it's returned to you. We return this value when the LocationCode is Returned.

gsoTrackingNumber :: Lens' GetStatusResponse (Maybe Text)

gsoSignatureFileContents :: Lens' GetStatusResponse (Maybe Text)

gsoSignature :: Lens' GetStatusResponse (Maybe Text)

gsoProgressMessage :: Lens' GetStatusResponse (Maybe Text)

gsoProgressCode :: Lens' GetStatusResponse (Maybe Text)

gsoLogKey :: Lens' GetStatusResponse (Maybe Text)

gsoLogBucket :: Lens' GetStatusResponse (Maybe Text)

gsoLocationMessage :: Lens' GetStatusResponse (Maybe Text)

gsoLocationCode :: Lens' GetStatusResponse (Maybe Text)

gsoJobType :: Lens' GetStatusResponse (Maybe JobType)

gsoJobId :: Lens' GetStatusResponse (Maybe Text)

gsoErrorCount :: Lens' GetStatusResponse (Maybe Integer)

gsoCurrentManifest :: Lens' GetStatusResponse (Maybe Text)

gsoCreationDate :: Lens' GetStatusResponse (Maybe ISO8601)

gsoCarrier :: Lens' GetStatusResponse (Maybe Text)

gsoAwsShippingAddress :: Lens' GetStatusResponse (Maybe Text)

listJobs :: ListJobs

Minimum specification for a ListJobs request.

data ListJobs

Constructors

ListJobs

Fields

_ljiMarker :: Maybe Text

Specifies the JOBID to start after when listing the jobs created with your account. AWS Import/Export lists your jobs in reverse chronological order. See MaxJobs.

_ljiMaxJobs :: Maybe Integer

Sets the maximum number of jobs returned in the response. If there are additional jobs that were not returned because MaxJobs was exceeded, the response contains &lt;IsTruncated&gt;true&lt;/IsTruncated&gt;. To return the additional jobs, see Marker.

ljiMaxJobs :: Lens' ListJobs (Maybe Integer)

ljiMarker :: Lens' ListJobs (Maybe Text)

data ListJobsResponse

Constructors

ListJobsResponse

Fields

_ljoIsTruncated :: Bool

Indicates whether the list of jobs was truncated. If true, then call ListJobs again using the last JobId element as the marker.

_ljoJobs :: [Job]

A list container for Jobs returned by the ListJobs operation.

ljoJobs :: Lens' ListJobsResponse [Job]

ljoIsTruncated :: Lens' ListJobsResponse Bool

data ImportExport

Supported version (2010-06-01) of the AWS Import/Export service.

xmlOptions :: Tagged a XMLOptions

data JobType

Specifies whether the job to initiate is an import or export job.

Constructors

JobTypeExport
Export

JobTypeImport
Import

data Job

Representation of a job returned by the ListJobs operation.

Constructors

Job

Fields

_jJobType :: JobType

Specifies whether the job to initiate is an import or export job.

_jJobId :: Text

A unique identifier which refers to a particular job.

_jIsCanceled :: Bool

Indicates whether the job was canceled.

_jCreationDate :: ISO8601

Timestamp of the CreateJob request in ISO8601 date format. For example "2010-03-28T20:27:35Z".

jJobType :: Lens' Job JobType

jJobId :: Lens' Job Text

jIsCanceled :: Lens' Job Bool

jCreationDate :: Lens' Job ISO8601

data UpdateJob

Constructors

UpdateJob

Fields

_ujiJobId :: Text

A unique identifier which refers to a particular job.

_ujiJobType :: JobType

Specifies whether the job to initiate is an import or export job.

_ujiManifest :: Text

The UTF-8 encoded text of the manifest file.

_ujiValidateOnly :: Bool

Validate the manifest and parameter values in the request but do not actually create a job.

ujiValidateOnly :: Lens' UpdateJob Bool

ujiManifest :: Lens' UpdateJob Text

ujiJobType :: Lens' UpdateJob JobType

ujiJobId :: Lens' UpdateJob Text

data UpdateJobResponse

Constructors

UpdateJobResponse

Fields

_ujoSuccess :: Maybe Bool

Specifies whether (true) or not (false) AWS Import/Export updated your job.

_ujoWarningMessage :: Maybe Text

An optional message notifying you of non-fatal issues with the job, such as use of an incompatible Amazon S3 bucket name.

ujoWarningMessage :: Lens' UpdateJobResponse (Maybe Text)

ujoSuccess :: Lens' UpdateJobResponse (Maybe Bool)