Network.AWS.ImportExport.V2010_06_01.ListJobs

Description

This operation returns the jobs associated with the requester. AWS Import/Export lists the jobs in reverse chronological order based on the date of creation. For example if Job Test1 was created 2009Dec30 and Test2 was created 2010Feb05, the ListJobs operation would return Test2 followed by Test1.

Documentation

listJobs :: ListJobs

Minimum specification for a ListJobs request.

data ListJobs

Constructors

ListJobs

Fields

_ljiMarker :: Maybe Text

Specifies the JOBID to start after when listing the jobs created with your account. AWS Import/Export lists your jobs in reverse chronological order. See MaxJobs.

_ljiMaxJobs :: Maybe Integer

Sets the maximum number of jobs returned in the response. If there are additional jobs that were not returned because MaxJobs was exceeded, the response contains &lt;IsTruncated&gt;true&lt;/IsTruncated&gt;. To return the additional jobs, see Marker.

ljiMaxJobs :: Lens' ListJobs (Maybe Integer)

ljiMarker :: Lens' ListJobs (Maybe Text)

data ListJobsResponse

Constructors

ListJobsResponse

Fields

_ljoIsTruncated :: Bool

Indicates whether the list of jobs was truncated. If true, then call ListJobs again using the last JobId element as the marker.

_ljoJobs :: [Job]

A list container for Jobs returned by the ListJobs operation.

ljoJobs :: Lens' ListJobsResponse [Job]

ljoIsTruncated :: Lens' ListJobsResponse Bool