Network.AWS.ImportExport.V2010_06_01.Types

Description

AWS Import/Export accelerates moving large amounts of data into and out of AWS using portable storage devices for transport. AWS transfers your data directly onto and off of storage devices using Amazon’s high-speed internal network and bypassing the Internet. For significant data sets, AWS Import/Export is often faster than Internet transfer and more cost effective than upgrading your connectivity.

Documentation

data ImportExport

Supported version (2010-06-01) of the AWS Import/Export service.

xmlOptions :: Tagged a XMLOptions

data JobType

Specifies whether the job to initiate is an import or export job.

Constructors

JobTypeExport
Export

JobTypeImport
Import

data Job

Representation of a job returned by the ListJobs operation.

Constructors

Job

Fields

_jJobType :: JobType

Specifies whether the job to initiate is an import or export job.

_jJobId :: Text

A unique identifier which refers to a particular job.

_jIsCanceled :: Bool

Indicates whether the job was canceled.

_jCreationDate :: ISO8601

Timestamp of the CreateJob request in ISO8601 date format. For example "2010-03-28T20:27:35Z".

jJobType :: Lens' Job JobType

jJobId :: Lens' Job Text

jIsCanceled :: Lens' Job Bool

jCreationDate :: Lens' Job ISO8601