Network.AWS.ImportExport.CreateJob

Description

This operation initiates the process of scheduling an upload or download of your data. You include in the request a manifest that describes the data transfer specifics. The response to the request includes a job ID, which you can use in other operations, a signature that you use to identify your storage device, and the address where you should ship your storage device.

Creating a Request

createJob

Arguments

:: JobType
cjJobType

-> Text
cjManifest

-> Bool
cjValidateOnly

-> CreateJob

Creates a value of CreateJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cjAPIVersion - Undocumented member.

cjManifestAddendum - Undocumented member.

cjJobType - Undocumented member.

cjManifest - Undocumented member.

cjValidateOnly - Undocumented member.

data CreateJob

Input structure for the CreateJob operation.

See: createJob smart constructor.

Request Lenses

cjAPIVersion :: Lens' CreateJob (Maybe Text)

Undocumented member.

cjManifestAddendum :: Lens' CreateJob (Maybe Text)

Undocumented member.

cjJobType :: Lens' CreateJob JobType

Undocumented member.

cjManifest :: Lens' CreateJob Text

Undocumented member.

cjValidateOnly :: Lens' CreateJob Bool

Undocumented member.

Destructuring the Response

createJobResponse

Arguments

:: Int
cjrsResponseStatus

-> CreateJobResponse

Creates a value of CreateJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cjrsSignature - Undocumented member.

cjrsJobType - Undocumented member.

cjrsJobId - Undocumented member.

cjrsSignatureFileContents - Undocumented member.

cjrsWarningMessage - Undocumented member.

cjrsArtifactList - Undocumented member.

cjrsResponseStatus - -- | The response status code.

data CreateJobResponse

Output structure for the CreateJob operation.

See: createJobResponse smart constructor.

Response Lenses

cjrsSignature :: Lens' CreateJobResponse (Maybe Text)

Undocumented member.

cjrsJobType :: Lens' CreateJobResponse (Maybe JobType)

Undocumented member.

cjrsJobId :: Lens' CreateJobResponse (Maybe Text)

Undocumented member.

cjrsSignatureFileContents :: Lens' CreateJobResponse (Maybe Text)

Undocumented member.

cjrsWarningMessage :: Lens' CreateJobResponse (Maybe Text)

Undocumented member.

cjrsArtifactList :: Lens' CreateJobResponse [Artifact]

Undocumented member.

cjrsResponseStatus :: Lens' CreateJobResponse Int

- | The response status code.