Network.AWS.ImportExport.CreateJob

Description

This operation initiates the process of scheduling an upload or download of your data. You include in the request a manifest that describes the data transfer specifics. The response to the request includes a job ID, which you can use in other operations, a signature that you use to identify your storage device, and the address where you should ship your storage device.

http://docs.aws.amazon.com/AWSImportExport/latest/DG/WebCreateJob.html

Request

data CreateJob

Request constructor

createJob

Arguments

:: JobType
cjJobType

-> Text
cjManifest

-> Bool
cjValidateOnly

-> CreateJob

CreateJob constructor.

The fields accessible through corresponding lenses are:

cjAPIVersion :: Maybe Text

cjJobType :: JobType

cjManifest :: Text

cjManifestAddendum :: Maybe Text

cjValidateOnly :: Bool

Request lenses

cjAPIVersion :: Lens' CreateJob (Maybe Text)

cjJobType :: Lens' CreateJob JobType

cjManifest :: Lens' CreateJob Text

cjManifestAddendum :: Lens' CreateJob (Maybe Text)

cjValidateOnly :: Lens' CreateJob Bool

Response

data CreateJobResponse

Response constructor

createJobResponse :: CreateJobResponse

CreateJobResponse constructor.

The fields accessible through corresponding lenses are:

cjrArtifactList :: [Artifact]

cjrJobId :: Maybe Text

cjrJobType :: Maybe JobType

cjrSignature :: Maybe Text

cjrSignatureFileContents :: Maybe Text

cjrWarningMessage :: Maybe Text

Response lenses

cjrArtifactList :: Lens' CreateJobResponse [Artifact]

cjrJobId :: Lens' CreateJobResponse (Maybe Text)

cjrJobType :: Lens' CreateJobResponse (Maybe JobType)

cjrSignature :: Lens' CreateJobResponse (Maybe Text)

cjrSignatureFileContents :: Lens' CreateJobResponse (Maybe Text)

cjrWarningMessage :: Lens' CreateJobResponse (Maybe Text)