Network.AWS.ImportExport.GetStatus

Description

This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job. You can only return information about jobs you own.

Creating a Request

getStatus

Arguments

:: Text
gsJobId

-> GetStatus

Creates a value of GetStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsAPIVersion - Undocumented member.

gsJobId - Undocumented member.

data GetStatus

Input structure for the GetStatus operation.

See: getStatus smart constructor.

Request Lenses

gsAPIVersion :: Lens' GetStatus (Maybe Text)

Undocumented member.

gsJobId :: Lens' GetStatus Text

Undocumented member.

Destructuring the Response

getStatusResponse

Arguments

:: Int
gsrsResponseStatus

-> GetStatusResponse

Creates a value of GetStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsrsCarrier - Undocumented member.

gsrsTrackingNumber - Undocumented member.

gsrsSignature - Undocumented member.

gsrsJobType - Undocumented member.

gsrsJobId - Undocumented member.

gsrsSignatureFileContents - Undocumented member.

gsrsErrorCount - Undocumented member.

gsrsCurrentManifest - Undocumented member.

gsrsArtifactList - Undocumented member.

gsrsLogBucket - Undocumented member.

gsrsCreationDate - Undocumented member.

gsrsProgressCode - Undocumented member.

gsrsLocationCode - Undocumented member.

gsrsLogKey - Undocumented member.

gsrsLocationMessage - Undocumented member.

gsrsProgressMessage - Undocumented member.

gsrsResponseStatus - -- | The response status code.

data GetStatusResponse

Output structure for the GetStatus operation.

See: getStatusResponse smart constructor.

Response Lenses

gsrsCarrier :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsTrackingNumber :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsSignature :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsJobType :: Lens' GetStatusResponse (Maybe JobType)

Undocumented member.

gsrsJobId :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsSignatureFileContents :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsErrorCount :: Lens' GetStatusResponse (Maybe Int)

Undocumented member.

gsrsCurrentManifest :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsArtifactList :: Lens' GetStatusResponse [Artifact]

Undocumented member.

gsrsLogBucket :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsCreationDate :: Lens' GetStatusResponse (Maybe UTCTime)

Undocumented member.

gsrsProgressCode :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsLocationCode :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsLogKey :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsLocationMessage :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsProgressMessage :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

gsrsResponseStatus :: Lens' GetStatusResponse Int

- | The response status code.