Network.AWS.ImportExport.Types

Description

Service Configuration

importExport :: Service

API version 2010-06-01 of the Amazon Import/Export SDK configuration.

Errors

_InvalidJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

The JOBID was missing, not found, or not associated with the AWS account.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters had an invalid value.

_ExpiredJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the specified job has expired out of the system.

_InvalidFileSystemException :: AsError a => Getting (First ServiceError) a ServiceError

File system specified in export manifest is invalid.

_InvalidAccessKeyIdException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Access Key ID specified in the request did not match the manifest's accessKeyId value. The manifest and the request authentication must use the same AWS Access Key ID.

_UnableToUpdateJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Import/Export cannot update the job

_UnableToCancelJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Import/Export cannot cancel the job

_MultipleRegionsException :: AsError a => Getting (First ServiceError) a ServiceError

Your manifest file contained buckets from multiple regions. A job is restricted to buckets from one region. Please correct and resubmit.

_InvalidVersionException :: AsError a => Getting (First ServiceError) a ServiceError

The client tool version is invalid.

_MalformedManifestException :: AsError a => Getting (First ServiceError) a ServiceError

Your manifest is not well-formed.

_MissingParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more required parameters was missing from the request.

_CanceledJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

The specified job ID has been canceled and is no longer valid.

_BucketPermissionException :: AsError a => Getting (First ServiceError) a ServiceError

The account specified does not have the appropriate bucket permissions.

_NoSuchBucketException :: AsError a => Getting (First ServiceError) a ServiceError

The specified bucket does not exist. Create the specified bucket or change the manifest's bucket, exportBucket, or logBucket field to a bucket that the account, as specified by the manifest's Access Key ID, has write permissions to.

_InvalidAddressException :: AsError a => Getting (First ServiceError) a ServiceError

The address specified in the manifest is invalid.

_MissingCustomsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more required customs parameters was missing from the manifest.

_InvalidManifestFieldException :: AsError a => Getting (First ServiceError) a ServiceError

One or more manifest fields was invalid. Please correct and resubmit.

_InvalidCustomsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more customs parameters was invalid. Please correct and resubmit.

_MissingManifestFieldException :: AsError a => Getting (First ServiceError) a ServiceError

One or more required fields were missing from the manifest file. Please correct and resubmit.

_CreateJobQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Each account can create only a certain number of jobs per day. If you need to create more than this, please contact awsimportexport@amazon.com to explain your particular use case.

JobType

data JobType

Specifies whether the job to initiate is an import or export job.

Constructors

Export

Import

Artifact

data Artifact

A discrete item that contains the description and URL of an artifact (such as a PDF).

See: artifact smart constructor.

artifact :: Artifact

Creates a value of Artifact with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aURL - Undocumented member.

aDescription - Undocumented member.

aURL :: Lens' Artifact (Maybe Text)

Undocumented member.

aDescription :: Lens' Artifact (Maybe Text)

Undocumented member.

Job

data Job

Representation of a job returned by the ListJobs operation.

See: job smart constructor.

job

Arguments

:: JobType
jobJobType

-> Text
jobJobId

-> Bool
jobIsCanceled

-> UTCTime
jobCreationDate

-> Job

Creates a value of Job with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jobJobType - Undocumented member.

jobJobId - Undocumented member.

jobIsCanceled - Undocumented member.

jobCreationDate - Undocumented member.

jobJobType :: Lens' Job JobType

Undocumented member.

jobJobId :: Lens' Job Text

Undocumented member.

jobIsCanceled :: Lens' Job Bool

Undocumented member.

jobCreationDate :: Lens' Job UTCTime

Undocumented member.