Network.AWS.ImportExport.UpdateJob

Description

You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file. The manifest file attached to this request replaces the original manifest file. You can only use the operation after a CreateJob request but before the data transfer starts and you can only use it on jobs you own.

http://docs.aws.amazon.com/AWSImportExport/latest/DG/WebUpdateJob.html

Request

data UpdateJob

Request constructor

updateJob

Arguments

:: Text
ujJobId

-> Text
ujManifest

-> JobType
ujJobType

-> Bool
ujValidateOnly

-> UpdateJob

UpdateJob constructor.

The fields accessible through corresponding lenses are:

ujAPIVersion :: Maybe Text

ujJobId :: Text

ujJobType :: JobType

ujManifest :: Text

ujValidateOnly :: Bool

Request lenses

ujAPIVersion :: Lens' UpdateJob (Maybe Text)

ujJobId :: Lens' UpdateJob Text

ujJobType :: Lens' UpdateJob JobType

ujManifest :: Lens' UpdateJob Text

ujValidateOnly :: Lens' UpdateJob Bool

Response

data UpdateJobResponse

Response constructor

updateJobResponse :: UpdateJobResponse

UpdateJobResponse constructor.

The fields accessible through corresponding lenses are:

ujrArtifactList :: [Artifact]

ujrSuccess :: Maybe Bool

ujrWarningMessage :: Maybe Text

Response lenses

ujrArtifactList :: Lens' UpdateJobResponse [Artifact]

ujrSuccess :: Lens' UpdateJobResponse (Maybe Bool)

ujrWarningMessage :: Lens' UpdateJobResponse (Maybe Text)