Network.AWS.ImportExport.V2010_06_01.ListJobs

Description

This operation returns the jobs associated with the requester. AWS Import/Export lists the jobs in reverse chronological order based on the date of creation. For example if Job Test1 was created 2009Dec30 and Test2 was created 2010Feb05, the ListJobs operation would return Test2 followed by Test1.

Request

data ListJobs

Input structure for the ListJobs operation.

Request constructor

mkListJobs :: ListJobs

Smart constructor for the minimum required parameters to construct a valid ListJobs request.

Request lenses

ljMaxJobs :: Lens' ListJobs (Maybe Integer)

Sets the maximum number of jobs returned in the response. If there are additional jobs that were not returned because MaxJobs was exceeded, the response contains &lt;IsTruncated&gt;true&lt;/IsTruncated&gt;. To return the additional jobs, see Marker.

ljMarker :: Lens' ListJobs (Maybe Text)

Specifies the JOBID to start after when listing the jobs created with your account. AWS Import/Export lists your jobs in reverse chronological order. See MaxJobs.

Response

data ListJobsResponse

Output structure for the ListJobs operation.

Response constructor

mkListJobsResponse

Arguments

:: Bool
ljrIsTruncated

-> ListJobsResponse

Smart constructor for the minimum required parameters to construct a valid ListJobsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

ljrJobs :: Lens' ListJobsResponse [Job]

A list container for Jobs returned by the ListJobs operation.

ljrIsTruncated :: Lens' ListJobsResponse Bool

Indicates whether the list of jobs was truncated. If true, then call ListJobs again using the last JobId element as the marker.