Network.AWS.ImportExport.V2010_06_01.Types

Description

AWS Import/Export accelerates moving large amounts of data into and out of AWS using portable storage devices for transport. AWS transfers your data directly onto and off of storage devices using Amazon’s high-speed internal network and bypassing the Internet. For significant data sets, AWS Import/Export is often faster than Internet transfer and more cost effective than upgrading your connectivity.

Service

data ImportExport

Supported version (2010-06-01) of the AWS Import/Export service.

XML

xmlOptions :: Tagged a XMLOptions

JobType

data JobType

Constructors

JobTypeExport
Export

JobTypeImport
Import

Job

data Job

Representation of a job returned by the ListJobs operation.

mkJob

Arguments

:: Text
jJobId

-> ISO8601
jCreationDate

-> Bool
jIsCanceled

-> JobType
jJobType

-> Job

Smart constructor for the minimum required fields to construct a valid Job data type.

Job is exclusively used in responses and this constructor is provided for convenience and testing purposes.

jJobId :: Lens' Job Text

A unique identifier which refers to a particular job.

jCreationDate :: Lens' Job ISO8601

Timestamp of the CreateJob request in ISO8601 date format. For example "2010-03-28T20:27:35Z".

jIsCanceled :: Lens' Job Bool

Indicates whether the job was canceled.

jJobType :: Lens' Job JobType

Specifies whether the job to initiate is an import or export job.